cloud_function/requirements.txt
google-resumable-media==0.6.0
google-cloud-storage==1.30.0
google-cloud-bigquery==1.26.1
pandas
numpy==1.18.1
tensorflow==2.7.0
tensorflow_hub==0.12.0
pyparsing==2.4.2